Refactor: Modernize TOL codebase for Linux compatibility
This commit modernizes the TOL (Time-Oriented Language) codebase to build
successfully on modern Linux systems with GCC and recent C++ standards.

Major changes include:

## Hash Map Compatibility
- Replace MSVC-specific hash maps with Google hash maps for Linux
- Fix hash_value function definitions to be MSVC-specific only
- Update hash map includes and conditional compilation

## Header and Include Fixes
- Add missing <climits> includes for INT_MAX/LONG_MAX constants
- Fix integer overflow warnings by using INT_MAX instead of LONG_MAX
- Resolve CBLAS header conflicts between GSL and system libraries

## Platform-Specific Code Fixes
- Fix conditional compilation for Windows-only headers (windows.h, dos.h)
- Update platform detection macros for better Linux compatibility
- Make Windows-specific dynamic library loading conditional

## String Literal Compatibility
- Fix C++11 string literal concatenation issues in matrix types
- Update macro definitions to properly concatenate string literals
- Resolve user-defined string literal operator conflicts

## Memory Management System
- Add proper includes for BFixedSizeMemoryBase system
- Fix memory management header dependencies

## Build System Improvements
- Update CMake configuration for modern Linux systems
- Fix library linking and dependency resolution

Progress: ~80% of modules now build successfully on Linux.
Remaining work involves completing memory management system fixes
and resolving final string literal issues in serie_type modules.

Tested on: Ubuntu with GCC, CMake 3.x, GSL, BLAS/LAPACK libraries

+++ tol-master/tol/btol/matrix_type/tol_mat.hpp
@@ -27,6 +27,7 @@
 
 #if GCC_VERSION >= 40300
 #include <limits>
+#include <climits>
 #endif
 
 //--------------------------------------------------------------------
@@ -129,7 +130,7 @@ void BMatrix<Any>::Alloc(BInt r, BInt c)
  */
 //--------------------------------------------------------------------
 {
-  static int max_size = LONG_MAX;
+  static int max_size = INT_MAX;
   int i,j;
  
   if (rows_ == r && columns_ == c)
